# openapi.json
{
    "components": {
        "schemas": {
...
            "Box2DDataInput": {
                "properties": {
                    "height": {
                        "title": "Height",
                        "type": "number"
                    },
                    "rotation": {
                        "default": 0,
                        "title": "Rotation",
                        "type": "number"
                    },
                    "width": {
                        "title": "Width",
                        "type": "number"
                    },
                    "x": {
                        "title": "X",
                        "type": "number"
                    },
                    "y": {
                        "title": "Y",
                        "type": "number"
                    }
                },
                "required": [
                    "x",
                    "y",
                    "width",
                    "height"
                ],
                "title": "Box2DDataInput",
                "type": "object"
            },
...
            "Keypoint2DDataInput": {
                "properties": {
                    "points": {
                        "items": {
                            "items": {
                                "type": "number"
                            },
                            "type": "array"
                        },
                        "title": "Points",
                        "type": "array"
                    }
                },
                "required": [
                    "points"
                ],
                "title": "Keypoint2DDataInput",
                "type": "object"
            },
...
            "PaginatedResponse_ProjectImageOutput_": {
                "properties": {
                    "count": {
                        "title": "Count",
                        "type": "integer"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectImageOutput"
                        },
                        "title": "Items",
                        "type": "array"
                    },
                    "limit": {
                        "title": "Limit",
                        "type": "integer"
                    },
                    "offset": {
                        "title": "Offset",
                        "type": "integer"
                    }
                },
                "required": [
                    "count",
                    "limit",
                    "offset",
                    "items"
                ],
                "title": "PaginatedResponse[ProjectImageOutput]",
                "type": "object"
            },
...
            "Polygon2DDataInput": {
                "properties": {
                    "points": {
                        "items": {
                            "items": {
                                "type": "number"
                            },
                            "type": "array"
                        },
                        "title": "Points",
                        "type": "array"
                    }
                },
                "required": [
                    "points"
                ],
                "title": "Polygon2DDataInput",
                "type": "object"
            },
...
            "ProjectAnnotationBatchInput": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectAnnotationItemInput"
                        },
                        "title": "Items",
                        "type": "array"
                    }
                },
                "required": [
                    "items"
                ],
                "title": "ProjectAnnotationBatchInput",
                "type": "object"
            },
            "ProjectAnnotationBatchOutput": {
                "properties": {
                    "created": {
                        "title": "Created",
                        "type": "integer"
                    },
                    "failed": {
                        "title": "Failed",
                        "type": "integer"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectAnnotationResultItem"
                        },
                        "title": "Results",
                        "type": "array"
                    }
                },
                "required": [
                    "created",
                    "failed",
                    "results"
                ],
                "title": "ProjectAnnotationBatchOutput",
                "type": "object"
            },
            "ProjectAnnotationItemInput": {
                "properties": {
                    "annotation_type": {
                        "title": "Annotation Type",
                        "type": "string"
                    },
                    "box": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Box2DDataInput"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "client_ref": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Ref"
                    },
                    "image_id": {
                        "title": "Image Id",
                        "type": "integer"
                    },
                    "keypoint": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Keypoint2DDataInput"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "label": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Label"
                    },
                    "polygon": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Polygon2DDataInput"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "required": [
                    "image_id",
                    "annotation_type"
                ],
                "title": "ProjectAnnotationItemInput",
                "type": "object"
            },
            "ProjectAnnotationResultItem": {
                "properties": {
                    "annotation_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Annotation Id"
                    },
                    "client_ref": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Ref"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    },
                    "image_id": {
                        "title": "Image Id",
                        "type": "integer"
                    },
                    "status": {
                        "title": "Status",
                        "type": "string"
                    }
                },
                "required": [
                    "image_id",
                    "status"
                ],
                "title": "ProjectAnnotationResultItem",
                "type": "object"
            },
...
            "ProjectImageOutput": {
                "properties": {
                    "file_name": {
                        "title": "File Name",
                        "type": "string"
                    },
                    "file_url": {
                        "title": "File Url",
                        "type": "string"
                    },
                    "height": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Height"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "width": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Width"
                    }
                },
                "required": [
                    "id",
                    "file_name",
                    "width",
                    "height",
                    "file_url"
                ],
                "title": "ProjectImageOutput",
                "type": "object"
            },
...
            "ProjectMetaOutput": {
                "properties": {
                    "created_at": {
                        "format": "date-time",
                        "title": "Created At",
                        "type": "string"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "label_mapping": {
                        "additionalProperties": true,
                        "title": "Label Mapping",
                        "type": "object"
                    },
                    "meta_info": {
                        "additionalProperties": true,
                        "title": "Meta Info",
                        "type": "object"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "title": "Updated At",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description",
                    "meta_info",
                    "label_mapping",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProjectMetaOutput",
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWTAuth": {
                "scheme": "bearer",
                "type": "http"
            },
            "ProjectAPIKeyAuth": {
                "in": "header",
                "name": "X-API-Key",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "",
        "title": "Anno API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
        "/api/infers/project/annotations": {
            "post": {
                "operationId": "projectinfer_submit_annotations_4b380a1f",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectAnnotationBatchInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectAnnotationBatchOutput"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "ProjectAPIKeyAuth": []
                    }
                ],
                "summary": "Submit Annotations",
                "tags": [
                    "infer-project"
                ]
            }
        },
        "/api/infers/project/images": {
            "get": {
                "operationId": "projectinfer_list_images_caf9b5dc",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 100,
                            "title": "Limit",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "required": false,
                        "schema": {
                            "default": 0,
                            "title": "Offset",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "has_active_annotations",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Has Active Annotations"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_ProjectImageOutput_"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "ProjectAPIKeyAuth": []
                    }
                ],
                "summary": "List Images",
                "tags": [
                    "infer-project"
                ]
            }
        },
...
        "/api/infers/project/meta": {
            "get": {
                "operationId": "projectinfer_project_meta_bfb1d6bf",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectMetaOutput"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "ProjectAPIKeyAuth": []
                    }
                ],
                "summary": "Project Meta",
                "tags": [
                    "infer-project"
                ]
            }
        },
...
    },
    "servers": []
}